refactoring the name of the jobTitle dto model
