+++ app/src/main/java/com/cupcake/jobsfinder/data/dto/JobTitleDto.kt
@@ -1,4 +1,4 @@
-package com.cupcake.jobsfinder.data.model
+package com.cupcake.jobsfinder.data.dto
 
 data class JobTitleDto(
     val id: Long,
