+++ app/src/main/java/com/cupcake/jobsfinder/data/repository/Repository.kt
@@ -1,6 +1,6 @@
 package com.cupcake.jobsfinder.data.repository
 
-import com.cupcake.jobsfinder.data.model.JobTitleDto
+import com.cupcake.jobsfinder.data.dto.JobTitleDto
 
 interface Repository {
     suspend fun JobTitle():List<JobTitleDto>
