+++ app/src/main/java/com/cupcake/jobsfinder/data/repository/RepositoryImpl.kt
@@ -1,6 +1,6 @@
 package com.cupcake.jobsfinder.data.repository
 
-import com.cupcake.jobsfinder.data.model.JobTitleDto
+import com.cupcake.jobsfinder.data.dto.JobTitleDto
 
 class RepositoryImpl : Repository {
     override suspend fun JobTitle(): List<JobTitleDto> {
