+++ app/src/main/java/com/cupcake/jobsfinder/domain/mapper/JobTitleMapper.kt
@@ -1,6 +1,7 @@
 package com.cupcake.jobsfinder.domain.mapper
 
-import com.cupcake.jobsfinder.data.model.JobTitleDto
+import com.cupcake.jobsfinder.data.dto.JobTitleDto
+import com.cupcake.jobsfinder.domain.model.JobTitle
 import com.cupcake.jobsfinder.ui.mapper.Mapper
